Moved DB, Auth, types to react sdk
